requirements.txt
rpaframework-recognition
rpaframework
robocorp-windows
robocorp
robotframework-mainframe3270
getgauge  